requirements.txt
Flask
python-dotenv
google-generativeai
torch
diffusers
Pillow
flask-ngrok
pyngrok
requests